+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ### Features
 
 - [#33](https://github.com/wallix/awless/issues/33): Ability to set AWS profile using `aws.profile` config key
+
+### Bugfixes
+
+- [#44](https://github.com/wallix/awless/issues/44): Fetch only the S3 buckets and related objects of the current region.